TotalDistanceComponent complete
