+++ MiniChallenge05/MiniChallenge05 Watch App/Components/TotalDistanceComponent.swift
@@ -1,0 +1,47 @@
+//
+//  TotalDistanceView.swift
+//  MiniChallenge05 Watch App
+//
+//  Created by Guilherme Nunes Lobo on 23/05/24.
+//
+
+import SwiftUI
+
+struct TotalDistanceComponent: View {
+    @StateObject private var healthKitManager = HealthKitManager(model: ModelNew())
+    @State private var totalDistance: Double = 0
+
+    var body: some View {
+        VStack {
+            HStack {
+                Image(systemName: "figure.walk")
+                    .font(.system(size: 30))
+                Text("0,\(Int(totalDistance * 10) % 10) KM")
+                    .font(.system(size: 30, weight: .regular, design: .default))
+            }
+            .padding(.bottom)
+
+            Text("CAMINHADOS")
+                .font(.system(size: 25, weight: .regular, design: .default))
+                .foregroundColor(.primary)
+
+            Spacer()
+        }
+        .padding()
+        .onAppear {
+            fetchTotalDistance()
+        }
+    }
+
+    private func fetchTotalDistance() {
+        healthKitManager.fetchTotalWalkingRunningDistance { distance in
+            DispatchQueue.main.async {
+                self.totalDistance = distance / 1000
+            }
+        }
+    }
+}
+
+#Preview {
+    TotalDistanceComponent()
+}
