+++ MiniChallenge05/MiniChallenge05 Watch App/MiniChallenge05App.swift
@@ -13,7 +13,7 @@ struct MiniChallenge05_Watch_AppApp: App {
     
     var body: some Scene {
         WindowGroup {
-            TotalDistanceView()
+            ContentView()
         }
         .modelContainer(for: ModelNew.self)
     }
